tag_for_variant follow-ups
